+++ spec/integration/inputs/kafka_spec.rb
@@ -79,6 +79,7 @@ describe "inputs/kafka", :integration => true do
     producer = org.apache.kafka.clients.producer.KafkaProducer.new(props)
 
     producer.send(record)
+    producer.flush
     producer.close
   end
 
@@ -185,9 +186,104 @@ describe "inputs/kafka", :integration => true do
       end
     end
   end
+
+  context "static membership 'group.instance.id' setting" do
+    let(:base_config) do
+      {
+        "topics" => ["logstash_integration_static_membership_topic"],
+        "group_id" => "logstash",
+        "consumer_threads" => 1,
+        # this is needed because the worker thread could be executed little after the producer sent the "up" message
+        "auto_offset_reset" => "earliest",
+        "group_instance_id" => "test_static_group_id"
+      }
+    end
+    let(:consumer_config) { base_config }
+    let(:logger) { double("logger") }
+    let(:queue) { java.util.concurrent.ArrayBlockingQueue.new(10) }
+    let(:kafka_input) { LogStash::Inputs::Kafka.new(consumer_config) }
+    before :each do
+      allow(LogStash::Inputs::Kafka).to receive(:logger).and_return(logger)
+      [:error, :warn, :info, :debug].each do |level|
+        allow(logger).to receive(level)
+      end
+
+      kafka_input.register
+    end
+
+    it "input plugin disconnects from the broker when another client with same static membership connects" do
+      expect(logger).to receive(:error).with("Another consumer with same group.instance.id has connected", anything)
+
+      input_worker = java.lang.Thread.new { kafka_input.run(queue) }
+      begin
+        input_worker.start
+        wait_kafka_input_is_ready("logstash_integration_static_membership_topic", queue)
+        saboteur_kafka_consumer = create_consumer_and_start_consuming("test_static_group_id")
+        saboteur_kafka_consumer.run # ask to be scheduled
+        saboteur_kafka_consumer.join
+
+        expect(saboteur_kafka_consumer.value).to eq("saboteur exited")
+      ensure
+        input_worker.join(30_000)
+      end
+    end
+
+    context "when the plugin is configured with multiple consumer threads" do
+      let(:consumer_config) { base_config.merge({"consumer_threads" => 2}) }
+
+      it "should avoid to connect with same 'group.instance.id'" do
+        expect(logger).to_not receive(:error).with("Another consumer with same group.instance.id has connected", anything)
+
+        input_worker = java.lang.Thread.new { kafka_input.run(queue) }
+        begin
+          input_worker.start
+          wait_kafka_input_is_ready("logstash_integration_static_membership_topic", queue)
+        ensure
+          kafka_input.stop
+          input_worker.join(1_000)
+        end
+      end
+    end
+  end
+end
+
+# return consumer Ruby Thread
+def create_consumer_and_start_consuming(static_group_id)
+  props = java.util.Properties.new
+  kafka = org.apache.kafka.clients.consumer.ConsumerConfig
+  props.put(kafka::BOOTSTRAP_SERVERS_CONFIG, "localhost:9092")
+  props.put(kafka::KEY_DESERIALIZER_CLASS_CONFIG, LogStash::Inputs::Kafka::DEFAULT_DESERIALIZER_CLASS)
+  props.put(kafka::VALUE_DESERIALIZER_CLASS_CONFIG, LogStash::Inputs::Kafka::DEFAULT_DESERIALIZER_CLASS)
+  props.put(kafka::GROUP_ID_CONFIG, "logstash")
+  props.put(kafka::GROUP_INSTANCE_ID_CONFIG, static_group_id)
+  consumer = org.apache.kafka.clients.consumer.KafkaConsumer.new(props)
+
+  Thread.new do
+    LogStash::Util::set_thread_name("integration_test_simple_consumer")
+    begin
+      consumer.subscribe(["logstash_integration_static_membership_topic"])
+      records = consumer.poll(java.time.Duration.ofSeconds(3))
+      "saboteur exited"
+    rescue => e
+      e # return the exception reached in thread.value
+    ensure
+      consumer.close
+    end
+  end
 end
 
 private
+
+def wait_kafka_input_is_ready(topic, queue)
+  # this is needed to give time to the kafka input to be up and running
+  header = org.apache.kafka.common.header.internals.RecordHeader.new("name", "Ping Up".to_java_bytes)
+  record = org.apache.kafka.clients.producer.ProducerRecord.new(topic, 0, "key", "value", [header])
+  send_message(record)
+
+  # Wait the message is processed
+  message = queue.poll(1, java.util.concurrent.TimeUnit::MINUTES)
+  expect(message).to_not eq(nil)
+end
 
 def consume_messages(config, queue: Queue.new, timeout:, event_count:)
   kafka_input = LogStash::Inputs::Kafka.new(config)
@@ -205,7 +301,7 @@ def consume_messages(config, queue: Queue.new, timeout:, event_count:)
 end
 
 
-describe "schema registry connection options" do
+describe "schema registry connection options", :integration => true do
   schema_registry = Manticore::Client.new
   before (:all) do
     shutdown_schema_registry
@@ -257,10 +353,13 @@ describe "schema registry connection options" do
   end
 end
 
-def save_avro_schema_to_schema_registry(schema_file, subject_name)
+def save_avro_schema_to_schema_registry(schema_file, subject_name, proto = 'http', port = 8081, manticore_options = {})
   raw_schema = File.readlines(schema_file).map(&:chomp).join
   raw_schema_quoted = raw_schema.gsub('"', '\"')
-  response = Manticore.post("http://localhost:8081/subjects/#{subject_name}/versions",
+
+  client = Manticore::Client.new(manticore_options)
+
+  response = client.post("#{proto}://localhost:#{port}/subjects/#{subject_name}/versions",
           body: '{"schema": "' + raw_schema_quoted + '"}',
           headers: {"Content-Type" => "application/vnd.schemaregistry.v1+json"})
   response
@@ -282,8 +381,17 @@ def startup_schema_registry(schema_registry, auth=false)
   end
 end
 
-describe "Schema registry API", :integration => true do
-  schema_registry = Manticore::Client.new
+shared_examples 'it has endpoints available to' do |tls|
+  let(:port) { tls ? 8083 : 8081 }
+  let(:proto) { tls ? 'https' : 'http' }
+
+  manticore_options = {
+    :ssl => {
+      :truststore => File.join(Dir.pwd, "tls_repository/clienttruststore.jks"),
+      :truststore_password => "changeit"
+    }
+  }
+  schema_registry = Manticore::Client.new(manticore_options)
 
   before(:all) do
     startup_schema_registry(schema_registry)
@@ -295,27 +403,38 @@ describe "Schema registry API", :integration => true do
 
   context 'listing subject on clean instance' do
     it "should return an empty set" do
-      subjects = JSON.parse schema_registry.get('http://localhost:8081/subjects').body
+      subjects = JSON.parse schema_registry.get("#{proto}://localhost:#{port}/subjects").body
       expect( subjects ).to be_empty
     end
   end
 
   context 'send a schema definition' do
     it "save the definition" do
-      response = save_avro_schema_to_schema_registry(File.join(Dir.pwd, "spec", "unit", "inputs", "avro_schema_fixture_payment.asvc"), "schema_test_1")
+      response = save_avro_schema_to_schema_registry(File.join(Dir.pwd, "spec", "unit", "inputs", "avro_schema_fixture_payment.asvc"), "schema_test_1", proto, port, manticore_options)
       expect( response.code ).to be(200)
       delete_remote_schema(schema_registry, "schema_test_1")
     end
 
     it "delete the schema just added" do
-      response = save_avro_schema_to_schema_registry(File.join(Dir.pwd, "spec", "unit", "inputs", "avro_schema_fixture_payment.asvc"), "schema_test_1")
+      response = save_avro_schema_to_schema_registry(File.join(Dir.pwd, "spec", "unit", "inputs", "avro_schema_fixture_payment.asvc"), "schema_test_1", proto, port, manticore_options)
       expect( response.code ).to be(200)
 
-      expect( schema_registry.delete('http://localhost:8081/subjects/schema_test_1?permanent=false').code ).to be(200)
+      expect( schema_registry.delete("#{proto}://localhost:#{port}/subjects/schema_test_1?permanent=false").code ).to be(200)
       sleep(1)
-      subjects = JSON.parse schema_registry.get('http://localhost:8081/subjects').body
+      subjects = JSON.parse schema_registry.get("#{proto}://localhost:#{port}/subjects").body
       expect( subjects ).to be_empty
     end
+  end
+end
+
+describe "Schema registry API", :integration => true do
+
+  context "when exposed with HTTPS" do
+    it_behaves_like 'it has endpoints available to', true
+  end
+
+  context "when exposed with plain HTTP" do
+    it_behaves_like 'it has endpoints available to', false
   end
 end
 
@@ -324,7 +443,13 @@ def shutdown_schema_registry
 end
 
 describe "Deserializing with the schema registry", :integration => true do
-  schema_registry = Manticore::Client.new
+  manticore_options = {
+    :ssl => {
+      :truststore => File.join(Dir.pwd, "tls_repository/clienttruststore.jks"),
+      :truststore_password => "changeit"
+    }
+  }
+  schema_registry = Manticore::Client.new(manticore_options)
 
   shared_examples 'it reads from a topic using a schema registry' do |with_auth|
 
@@ -423,28 +548,57 @@ describe "Deserializing with the schema registry", :integration => true do
     end
   end
 
-  context 'with an unauthed schema registry' do
+  shared_examples 'with an unauthed schema registry' do |tls|
+    let(:port) { tls ? 8083 : 8081 }
+    let(:proto) { tls ? 'https' : 'http' }
+
     let(:auth) { false }
     let(:avro_topic_name) { "topic_avro" }
-    let(:subject_url) { "http://localhost:8081/subjects" }
-    let(:plain_config)  { base_config.merge!({'schema_registry_url' => "http://localhost:8081"}) }
+    let(:subject_url) { "#{proto}://localhost:#{port}/subjects" }
+    let(:plain_config)  { base_config.merge!({
+      'schema_registry_url' => "#{proto}://localhost:#{port}",
+      'schema_registry_ssl_truststore_location' => File.join(Dir.pwd, "tls_repository/clienttruststore.jks"),
+      'schema_registry_ssl_truststore_password' => 'changeit',
+    }) }
 
     it_behaves_like 'it reads from a topic using a schema registry', false
   end
 
-  context 'with an authed schema registry' do
+  context 'with an unauthed schema registry' do
+    context "accessed through HTTPS" do
+      it_behaves_like 'with an unauthed schema registry', true
+    end
+
+    context "accessed through HTTPS" do
+      it_behaves_like 'with an unauthed schema registry', false
+    end
+  end
+
+  shared_examples 'with an authed schema registry' do |tls|
+    let(:port) { tls ? 8083 : 8081 }
+    let(:proto) { tls ? 'https' : 'http' }
     let(:auth) { true }
     let(:user) { "barney" }
     let(:password) { "changeme" }
     let(:avro_topic_name) { "topic_avro_auth" }
-    let(:subject_url) { "http://#{user}:#{password}@localhost:8081/subjects" }
+    let(:subject_url) { "#{proto}://#{user}:#{password}@localhost:#{port}/subjects" }
+    let(:tls_base_config) do
+      if tls
+        base_config.merge({
+          'schema_registry_ssl_truststore_location' => ::File.join(Dir.pwd, "tls_repository/clienttruststore.jks"),
+          'schema_registry_ssl_truststore_password' => 'changeit',
+        })
+      else
+        base_config
+      end
+    end
 
     context 'using schema_registry_key' do
       let(:plain_config) do
-        base_config.merge!({
-          'schema_registry_url' => "http://localhost:8081",
+        tls_base_config.merge!({
+          'schema_registry_url' => "#{proto}://localhost:#{port}",
           'schema_registry_key' => user,
-          'schema_registry_secret' => password
+          'schema_registry_secret' => password,
         })
       end
 
@@ -453,12 +607,22 @@ describe "Deserializing with the schema registry", :integration => true do
 
     context 'using schema_registry_url' do
       let(:plain_config) do
-        base_config.merge!({
-          'schema_registry_url' => "http://#{user}:#{password}@localhost:8081"
+        tls_base_config.merge!({
+          'schema_registry_url' => "#{proto}://#{user}:#{password}@localhost:#{port}",
         })
       end
 
       it_behaves_like 'it reads from a topic using a schema registry', true
+    end
+  end
+
+  context 'with an authed schema registry' do
+    context "accessed through HTTPS" do
+      it_behaves_like 'with an authed schema registry', true
+    end
+
+    context "accessed through HTTPS" do
+      it_behaves_like 'with an authed schema registry', false
     end
   end
 end